# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakLomba API Documentation",
        "description": "The **TernakLomba API** is a robust and scalable solution for managing competitive events, participants, scores, and results. \n *   It is designed to support a wide range of competitions, from small local events to large-scale tournaments. \n *  Key features include:\n *  - **Competition Management**: Create, update, and manage competitions with details such as name, description, start/end dates, and status.\n *  - **Participant Management**: Register and manage participants, including their details and competition assignments.\n *  - **Score Management**: Submit and manage scores for participants, with support for multiple judges and scoring criteria.\n *  - **Results and Leaderboards**: Automatically calculate and display competition results, including rankings and leaderboards.\n *  - **Judging Criteria**: Define custom judging criteria for competitions to ensure fair and transparent evaluations.\n *  - **Notifications**: Send real-time notifications to participants and judges about competition updates, scores, and results.\n *  This API is ideal for event organizers, educational institutions, and organizations hosting competitive events.",
        "contact": {
            "name": "the developer",
            "email": "support@ternaklomba.com"
        },
        "license": {
            "name": "Apache 2.0",
...
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/lombaCategory": {
            "get": {
                "tags": [
                    "Lomba Category"
                ],
                "summary": "Get All Lomba Category",
                "description": "Mengambil Semua data Category Lomba",
                "operationId": "lombaCategory-all",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengambil data category lomba",
                                    "data": [
                                        {
                                            "name": "University",
                                            "updated_at": "2025-03-20T17:11:42.000000Z",
                                            "created_at": "2025-03-20T17:11:42.000000Z",
                                            "id_lombaCategory": 1
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lomba Category"
                ],
                "summary": "Create Lomba Category",
                "description": "Membuat Category Lomba",
                "operationId": "lombaCategory-create",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Category Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "University"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil menambahkan data category lomba",
                                    "data": {
                                        "name": "University",
                                        "updated_at": "2025-03-20T17:11:42.000000Z",
                                        "created_at": "2025-03-20T17:11:42.000000Z",
                                        "id_lombaCategory": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lombaCategory/get/{id_lombaCategory}": {
            "get": {
                "tags": [
                    "Lomba Category"
                ],
                "summary": "Get One Lomba Category",
                "description": "Mengambil data Lomba Category",
                "operationId": "lombaCategory-get",
                "parameters": [
                    {
                        "name": "id_lombaCategory",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengambil data lomba category",
                                    "data": {
                                        "name": "University",
                                        "updated_at": "2025-03-20T17:11:42.000000Z",
                                        "created_at": "2025-03-20T17:11:42.000000Z",
                                        "id_lombaCategory": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/lombaCategory/{id_lombaCategory}": {
            "put": {
                "tags": [
                    "Lomba Category"
                ],
                "summary": "Update Lomba Category",
                "description": "Mengubah data Category Lomba",
                "operationId": "lombaCategory-update",
                "parameters": [
                    {
                        "name": "id_lombaCategory",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Category Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "University"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil mengubah data category lomba",
                                    "data": {
                                        "name": "University",
                                        "updated_at": "2025-03-20T17:11:42.000000Z",
                                        "created_at": "2025-03-20T17:11:42.000000Z",
                                        "id_lombaCategory": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lomba Category"
                ],
                "summary": "Delete Lomba Category",
                "description": "Menghapus data Category Lomba",
                "operationId": "lombaCategory-delete",
                "parameters": [
                    {
                        "name": "id_lombaCategory",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil menghapus data category lomba"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": false,
                                    "status_code": 401,
                                    "message": "Unauthorized"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}